# server/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Herbst MUD API",
    "description": "REST API for the Herbst MUD game server. All endpoints that modify data require JWT authentication via /users/auth.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/characters/{id}/skills": {
      "get": {
        "summary": "Get character skills (hotbar)",
        "responses": {},
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Characters"
        ]
      },
      "put": {
        "summary": "Set character skills",
        "responses": {},
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Characters"
        ]
      }
    },
...
    "/skills": {
      "get": {
        "summary": "List all skills",
        "responses": {},
        "tags": [
          "Skills"
        ]
      },
      "post": {
        "summary": "Create skill",
        "responses": {
          "201": {
            "description": "POST /skills",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skill"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillInput"
              }
            }
          }
        },
        "tags": [
          "Skills"
        ]
      }
    },
    "/skills/{id}": {
      "get": {
        "summary": "Get skill by ID",
        "responses": {
          "200": {
            "description": "GET /skills/{id}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skill"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Skills"
        ]
      },
      "put": {
        "summary": "Update skill",
        "responses": {
          "200": {
            "description": "PUT /skills/{id}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skill"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillInput"
              }
            }
          }
        },
        "tags": [
          "Skills"
        ]
      },
      "delete": {
        "summary": "Delete skill",
        "responses": {
          "204": {
            "description": "DELETE /skills/{id}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skill"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Skills"
        ]
      }
    },
...
    "/content/skills/{id}": {
      "get": {
        "summary": "Get content skill",
        "responses": {},
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Content"
        ]
      }
    },
...
    "/worlds/{world_id}/content/skills": {
      "get": {
        "summary": "List world skills",
        "responses": {},
        "parameters": [
          {
            "name": "world_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Worlds"
        ]
      }
    },
    "/worlds/{world_id}/content/skills/{skill_id}": {
      "get": {
        "summary": "Get world skill",
        "responses": {},
        "parameters": [
          {
            "name": "world_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Worlds"
        ]
      }
    },
...
    "/api/skills/{id}": {
      "put": {
        "summary": "Update skill faction fields",
        "responses": {
          "200": {
            "description": "PUT /api/skills/{id}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skill"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillUpdateInput"
              }
            }
          }
        },
        "tags": [
          "Skills"
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "SkillInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "skillType": {
            "type": "string"
          },
          "effectType": {
            "type": "string"
          },
          "effectValue": {
            "type": "number"
          },
          "cooldown": {
            "type": "integer"
          },
          "manaCost": {
            "type": "integer"
          },
          "staminaCost": {
            "type": "integer"
          }
        }
      },
...
      "Skill": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "skill_class": {
            "type": "string",
            "description": "passive or active"
          },
          "required_tag": {
            "type": "string",
            "description": "Tag required to unlock this skill beyond faction membership"
          },
          "proc_chance": {
            "type": "number",
            "format": "double",
            "description": "For passives: % chance to proc (0.15 = 15%)"
          },
          "proc_event": {
            "type": "string",
            "description": "What triggers the proc: on_hit, on_hit_received, on_crit, on_kill"
          },
          "cooldown_seconds": {
            "type": "integer",
            "description": "For actives: cooldown in seconds"
          }
        }
      },
      "SkillUpdateInput": {
        "type": "object",
        "properties": {
          "skill_class": {
            "type": "string"
          },
          "required_tag": {
            "type": "string"
          },
          "proc_chance": {
            "type": "number",
            "format": "double"
          },
          "proc_event": {
            "type": "string"
          },
          "cooldown_seconds": {
            "type": "integer"
          }
        }
      }
    }
  }
}